+++ CHANGES.md
@@ -6,3 +6,6 @@ information from a grid. see issue #3 and #4
 - fixed bugs occuring during storage as netCDF file see issue #8
 - comparison of grids is no longer using exact float comparison, see issue #9
 - added documentation and examples for working with the grid objects, see issue #1
+
+# #
+- added support for saving more subsets and loading a certain one in/from a netcdf grid file